Clear the leftovers the audit had been carrying
possible_values was silently ignored for integral arguments. The code was
wrong, not the docs: the same definition's constructor rejects being
handed both a list and a mapping, so it plainly meant to consult them.
SetValue checks both now, reporting InvalidArgument the way the string
path already did. The existing test only ever passed a value that was in
the list, which is why it survived.

Fixing that turned up something else: the builder cannot give an integral
argument a mapping at all, since that AddArgument overload is constrained
to enums and strings. The mapping branch is reachable only by building
the definition directly, which is what the new test does. Either the
constraint is too narrow or the constructor's check is pointless; the
question is recorded in the audit rather than guessed at.

New<T>'s null-allocator check becomes OPAL_VERIFY rather than the
OPAL_ASSERT the audit asked for, for the same reason as the rest of
section 10: an assert would leave release builds calling Alloc through a
null pointer.

The useless casts are gone. i64 is long long everywhere, so the LL suffix
already picks the right overload. A fourth site the audit had not listed
went with them.

CI gains a no-exceptions job on all three compilers. OPAL_EXCEPTIONS=OFF
is a different compile of every header, and nothing was checking it, so a
new throw would have broken that build silently. It builds the library
only: the option forces the tests off.

+++ test/json-reader-test.cpp
@@ -461,7 +461,7 @@ TEST_CASE("JsonReader i64 min", "[JsonReader]")
     JsonReader reader = ParseOrFail("-9223372036854775808");
     REQUIRE(reader.GetRoot().IsIntegerNumber());
     // Use explicit comparison to avoid literal issues.
-    REQUIRE(reader.GetRoot().GetIntegerNumber() == static_cast<i64>(-9223372036854775807LL - 1));
+    REQUIRE(reader.GetRoot().GetIntegerNumber() == (-9223372036854775807LL - 1));
 }
 
 TEST_CASE("JsonReader integer overflow falls back to f64", "[JsonReader]")
